+++ apps/nest-demo/src/app/text/text.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { CreateTextDto } from './dto/create-text.dto';
 import { UpdateTextDto } from './dto/update-text.dto';
-import * as request from 'request';
+import request from 'request';
 
 function include(name: string) {
   if (